# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: My Case commerce API
servers:
  - url: /api/v1
    description: Versioned commerce API
...
paths:
...
  /auth/telegram/verify:
    post:
      operationId: verifyTelegram
      tags: [auth]
      summary: Verify Telegram WebApp init data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramInitData"
      responses:
        "200":
          description: Verified Telegram identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramVerification"
  /telegram/link:
    post:
      operationId: linkTelegramCustomer
      tags: [customers]
      summary: Link the verified Telegram identity to a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramLinkInput"
      responses:
        "200":
          description: Linked customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramLink"
...
  /telegram/cart/items:
    post:
      operationId: upsertTelegramCartItem
      tags: [orders]
      summary: Add or replace a persistent cart line
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramCartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramCart"
  /telegram/cart/items/{productId}:
    delete:
      operationId: deleteTelegramCartItem
      tags: [orders]
      summary: Remove a persistent cart line
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramCart"
...
  /telegram/orders/{orderId}:
    get:
      operationId: getTelegramOrder
      tags: [orders]
      summary: Get one order belonging to the verified customer
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Customer order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramOrder"
components:
  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    TelegramInitData:
      type: object
      required: [init_data]
      properties:
        init_data: { type: string, minLength: 1, maxLength: 4096 }
    TelegramVerification:
      type: object
      required: [verified, telegram_user_id]
      properties:
        verified: { type: boolean }
        telegram_user_id: { type: string }
        username: { type: ["string", "null"] }
    TelegramLinkInput:
      type: object
      required: [full_name, phone]
      properties:
        full_name: { type: string, minLength: 1, maxLength: 160 }
        phone: { type: string, minLength: 5, maxLength: 40 }
        email: { type: ["string", "null"], maxLength: 255 }
    TelegramLink:
      type: object
      required: [linked, customer_id]
      properties:
        linked: { type: boolean }
        customer_id: { type: string }
    TelegramCartItemInput:
      type: object
      required: [product_id, quantity]
      properties:
        product_id: { type: string }
        quantity: { type: number, minimum: 1, maximum: 999 }
    TelegramCartItem:
      type: object
      required: [product_id, product_name, sku, quantity, unit_price, line_total]
      properties:
        product_id: { type: string }
        product_name: { type: string }
        sku: { type: string }
        quantity: { type: number }
        unit_price: { type: string }
        line_total: { type: string }
    TelegramCart:
      type: object
      required: [customer_id, items, total]
      properties:
        customer_id: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/TelegramCartItem"
        total: { type: string }
    TelegramOrderItem:
      type: object
      required: [product_id, product_name, sku, quantity, unit_price]
      properties:
        product_id: { type: string }
        product_name: { type: string }
        sku: { type: string }
        quantity: { type: number }
        unit_price: { type: string }
    TelegramOrder:
      type: object
      required: [id, status, total, created_at, items]
      properties:
        id: { type: string }
        status: { type: string }
        total: { type: string }
        created_at: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: "#/components/schemas/TelegramOrderItem"