# lib/api-zod/src/generated/types/telegramCartItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * My Case commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramCartItem {
  product_id: string;
  product_name: string;
  sku: string;
  quantity: number;
  unit_price: string;
  line_total: string;
}