# lib/api-zod/src/generated/types/telegramCartItemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * My Case commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramCartItemInput {
  product_id: string;
  /**
     * @minimum 1
     * @maximum 999
     */
  quantity: number;
}