# lib/api-zod/src/generated/types/telegramInitData.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * My Case commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramInitData {
  /**
     * @minLength 1
     * @maxLength 4096
     */
  init_data: string;
}