# lib/api-zod/src/generated/types/telegramLink.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * My Case commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramLink {
  linked: boolean;
  customer_id: string;
}