# lib/api-zod/src/generated/types/telegramLinkInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * My Case commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramLinkInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  full_name: string;
  /**
     * @minLength 5
     * @maxLength 40
     */
  phone: string;
  /**
     * @maxLength 255
     * @nullable
     */
  email?: string | null;
}