# lib/api-zod/src/generated/types/telegramOrder.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * My Case commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { TelegramOrderItem } from './telegramOrderItem';

export interface TelegramOrder {
  id: string;
  status: string;
  total: string;
  created_at: Date;
  items: TelegramOrderItem[];
}