# lib/api-zod/src/generated/types/telegramOrderItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * My Case commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramOrderItem {
  product_id: string;
  product_name: string;
  sku: string;
  quantity: number;
  unit_price: string;
}