# lib/api-zod/src/generated/types/telegramVerification.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * My Case commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramVerification {
  verified: boolean;
  telegram_user_id: string;
  /** @nullable */
  username?: string | null;
}